# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin-users]
      summary: List admin users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (staff cannot access)
    post:
      operationId: adminCreateUser
      tags: [admin-users]
      summary: Create an admin/moderator/staff user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin-users]
      summary: Update an admin user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: adminDeleteUser
      tags: [admin-users]
      summary: Delete an admin user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  schemas:
...
    AdminUser:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:        { type: string }
        email:     { type: string }
        name:      { type: string }
        role:      { type: string }
        createdAt: { type: string }
    AdminUserList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
    AdminUserInput:
      type: object
      required: [email, name, password, role]
      properties:
        email:    { type: string }
        name:     { type: string }
        password: { type: string }
        role:     { type: string }
    AdminUserUpdate:
      type: object
      properties:
        email:    { type: string }
        name:     { type: string }
        password: { type: string }
        role:     { type: string }
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
...